requirements.txt
discord.py
python-dotenv
aiohttp
pytz
vaderSentiment
openpipe